+++ CHANGELOG.md
@@ -6,5 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## 🔖 [0.1.0] - Unreleased
 ### ✨ Added
 - Unit test project using Catch2.
-- Adds `CLinqCollection<T>`
-    - With iterators.
+- Adds `CLinqCollection<T>` containing:
+    - `const` and non `const` iterators.
+    - Operators for comparisons, indexing and concatenation.